cleaned up view and js asset
